image_deblurMajor-master/requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
PyYAML>=5.4.1
tensorboardX>=2.2
tqdm>=4.50.0
Pillow>=8.0.0
scikit-image>=0.17.2
matplotlib>=3.3.0
opencv-python>=4.4.0
flask>=2.0.0
streamlit>=1.0.0
requests>=2.25.0
thop>=0.1.1
natsort>=8.0.0 